+++ spring-data-mongodb/src/main/java/org/springframework/data/mongodb/core/aggregation/ProjectionOperation.java
@@ -1395,7 +1395,7 @@ public class ProjectionOperation implements FieldsExposingAggregationOperation {
 				// implicit reference or explicit include?
 				if (value == null || Boolean.TRUE.equals(value)) {
 
-					if (Aggregation.SystemVariable.isReferingToSystemVariable(field.getTarget())) {
+					if (SystemVariable.isReferingToSystemVariable(field.getTarget())) {
 						return field.getTarget();
 					}
